requirements.txt
Keras == 2.2.0
Pillow == 5.1.0
h5py == 2.8.0
numpy == 1.14.5
opencv_python == 3.4.1.15
tensorflow == 1.8.0
